fix validation bug and inputMinigameData error

+++ src/main/java/de/unistuttgart/overworldbackend/data/PlayerTaskStatisticData.java
@@ -4,6 +4,8 @@ import java.util.UUID;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.PositiveOrZero;
+
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -24,7 +26,7 @@ public class PlayerTaskStatisticData {
 
   @Min(0)
   @Max(100)
-  long score;
+  int score;
 
   @NotNull
   String userId;
